fix: bigint to string

+++ src/app/api/territories-vs-residents/route.ts
@@ -3,13 +3,14 @@ import { query } from "@/lib/db";
 
 export async function GET() {
   const data = await query(
-    `SELECT d.nome_uf,
-       SUM(m.populacao_indigena_terra_indigena) AS em_terra_indigena,
-       SUM(m.populacao_indigena) AS total_indigena
-FROM municipio m
-JOIN diretorios_brasil_municipio d ON m.id_municipio = d.id_municipio
-GROUP BY d.nome_uf
-ORDER BY d.nome_uf;`
+    `SELECT 
+       d.nome_uf,
+       SUM(m.populacao_indigena_terra_indigena)::int AS em_terra_indigena,
+       SUM(m.populacao_indigena)::int AS total_indigena
+     FROM municipio m
+     JOIN diretorios_brasil_municipio d ON m.id_municipio = d.id_municipio
+     GROUP BY d.nome_uf
+     ORDER BY d.nome_uf;`
   );
   return NextResponse.json(data);
 }
